requirements_unbuntu.txt
asgiref==3.3.1
Django==3.1.3
django-filter
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
importlib-metadata==2.0.0
Markdown==3.3.3
python-dateutil==2.8.1
six==1.15.0
psycopg2-binary
PyJWT==1.7.1
django-oauth-toolkit==1.3.3
# edx-django-oauth2-provider==1.3.5

djangorestframework_simplejwt
django-cors-headers

django-background-tasks

python-dotenv==0.15.0

pandas==1.2.3
xlrd==2.0.1
openpyxl==3.0.7

tensorflow 
scipy
sklearn
